% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cprad.R
\name{cprad}
\alias{cprad}
\title{Generates CPRAd from vectors of unacceptable antigens}
\usage{
cprad(vector.A, vector.B, vector.Bw, vector.C, vector.Q, vector.R, decimals)
}
\arguments{
\item{Vector}{containing HLA-A antigens}

\item{Vector}{containing HLA-B antigens}

\item{Vector}{containing HLA-Bw antigens}

\item{Vector}{containing HLA-C antigens}

\item{Vector}{containing HLA-Q antigens}

\item{Vector}{containing HLA-R antigens}

\item{Number}{of decimal points}
}
\value{
CPRAd calculated from UNOS tables.
}
\description{
Generates CPRAd from vectors of unacceptable antigens
}
\examples{
\dontrun{
cprad<-function(vector.A,vector.B,vector.Bw,vector.C,vector.Q,vector.R,decimals)
}
}

