% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lcd.R
\name{lcd}
\alias{lcd}
\title{Generates CPRAd from vectors of unacceptable antigens}
\usage{
lcd(vector.A, vector.B, vector.Bw, vector.C, vector.Q, vector.R)
}
\arguments{
\item{Vector}{containing HLA-A antigens}

\item{Vector}{containing HLA-B antigens}

\item{Vector}{containing HLA-Bw antigens}

\item{Vector}{containing HLA-C antigens}

\item{Vector}{containing HLA-Q antigens}

\item{Vector}{containing HLA-R antigens}
}
\value{
Likelihood of a compatible donor.
}
\description{
Generates CPRAd from vectors of unacceptable antigens
}
\examples{
\dontrun{
lcd<-function(vector.A,vector.B,vector.Bw,vector.C,vector.Q,vector.R)
}
}

