% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pak_apply.R
\name{pak_apply}
\alias{pak_apply}
\title{Generates S1-S5 tables by pulling HLA-antigen frequency data for all antigens from tables.}
\usage{
pak_apply(vector.A, vector.B, vector.C, vector.Q, vector.R)
}
\arguments{
\item{Vector}{containing HLA-A antigens}

\item{Vector}{containing HLA-B antigens}

\item{Vector}{containing HLA-C antigens}

\item{Vector}{containing HLA-Q antigens}

\item{Vector}{containing HLA-R antigens}
}
\value{
List containing the S1-S5 tables.
}
\description{
Generates S1-S5 tables by pulling HLA-antigen frequency data for all antigens from tables.
}
\examples{
\dontrun{
pak_apply(A.excludes,B.excludes,C.excludes,Q.excludes,R.excludes)
}
}

