% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pak_calc.R
\name{pak_calc}
\alias{pak_calc}
\title{Calculates CPRAd from S1-S5 tables.}
\usage{
pak_calc(In, weights, decimals)
}
\arguments{
\item{List}{containing the S1-S5 tables}

\item{Vector}{of the ethnic weights.}

\item{Number}{of decimal points to use.}
}
\value{
CPRAd
}
\description{
Calculates CPRAd from S1-S5 tables.
}
\examples{
\dontrun{
pak_calc(in,weights,decimals)
}
}

