% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pak_equiv.R
\name{pak_equiv}
\alias{pak_equiv}
\title{Calculates HLA antigen equivalencies.}
\usage{
pak_equiv(in.A, in.B, in.Bw, in.C, in.Q, in.R)
}
\arguments{
\item{Vector}{containing HLA-A antigens}

\item{Vector}{containing HLA-B antigens}

\item{Vector}{containing HLA-Bw antigens}

\item{Vector}{containing HLA-C antigens}

\item{Vector}{containing HLA-Q antigens}

\item{Vector}{containing HLA-R antigens}
}
\value{
List of vectors, each containing the HLA-A, -B, -C, -Q, and -R antigens now with their equivalent antigens.
}
\description{
Calculates HLA antigen equivalencies.
}
\examples{
\dontrun{
pak_equiv(A.excludes,B.excludes,Bw.exludes,C.excludes,Q.excludes,R.excludes)
}
}

